server/requirements.txt
Django==4.1.4
django-cors-headers==3.7.0
djangorestframework==3.12.2
djangorestframework-simplejwt==5.2.2
numpy==1.24.2
pandas==1.5.3
scikit-learn==1.2.2
scipy==1.10.1
seaborn==0.12.2
psycopg2-binary==2.9.5
django-environ==0.10.0
djangorestframework==3.12.2
PyJWT==2.0.1